server/requirements-dev.txt
pytest
pytest-cov
pytest-flask
pytest-mock
requests
itsdangerous
pgpy
